Add Vercel deployment configuration
